requirements.txt
PyQt6==6.4.2
pyserial==3.5
pyqtgraph==0.13.3
